rename file containing ContainerRegistryContentClient
